app/requirements.txt
fastapi
uvicorn
pydantic
pillow